requirements.txt
sqlalchemy
requests
pyyaml
mysql.connector
mysqlclient

